Sectioning, content, style development.
Here content gets added in new categories, so that the page can act as a
hub and reference an eventual statement of work / portfolio page. Adding
the new content allows me to actually design content-out and refine the
content model. Flexbox has been applied in the relevant places and a
very minimal set of rules makes for a responsive multicolumn layout.